backend/requirements.txt
# Backend Dependencies - SecureCheck

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Environment Variables
python-dotenv==1.0.0

# HTTP Client (for security scans)
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Security Scanning Tools
python-nmap==0.7.1
validators==0.22.0

# Utilities
python-dateutil==2.8.2
loguru==0.7.2

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0